# vm: optimize eq, slt, sgt and iszero + tests