This PR adds testcases and benchmarks for a few very cheap and common vm ops. No rush to get it in for 1.8.0. 

On master (after adding tests):
```
go test . -run - -bench BenchmarkOpSg -benchmem -count=10 > oldsgt.txt
go test . -run - -bench BenchmarkOpIsZ -benchmem -count=10 > oldiszero.txt
go test . -run - -bench BenchmarkOpSl -benchmem -count=10 > oldslt.txt
go test . -run - -bench BenchmarkOpEq -benchmem -count=10 > oldeq.txt
``` 

With the changes
```
go test . -run - -bench BenchmarkOpSg -benchmem -count=10 > newsgt.txt
go test . -run - -bench BenchmarkOpIsZ -benchmem -count=10 > newiszero.txt
go test . -run - -bench BenchmarkOpSl -benchmem -count=10 > newslt.txt
go test . -run - -bench BenchmarkOpEq -benchmem -count=10 > neweq.txt
```

Result: 

### SGT

```
#~/go/bin/benchstat oldsgt.txt newsgt.txt 
name     old time/op    new time/op    delta
OpSgt-8     676ns ± 4%     392ns ± 7%  -41.96%  (p=0.000 n=10+10)

name     old alloc/op   new alloc/op   delta
OpSgt-8      448B ± 0%      192B ± 0%  -57.14%  (p=0.000 n=10+10)

name     old allocs/op  new allocs/op  delta
OpSgt-8      9.00 ± 0%      4.00 ± 0%  -55.56%  (p=0.000 n=10+10)
```

### SLT

```
name     old time/op    new time/op    delta
OpSlt-8     686ns ± 8%     392ns ±11%  -42.97%  (p=0.000 n=10+10)

name     old alloc/op   new alloc/op   delta
OpSlt-8      448B ± 0%      192B ± 0%  -57.14%  (p=0.000 n=10+10)

name     old allocs/op  new allocs/op  delta
OpSlt-8      9.00 ± 0%      4.00 ± 0%  -55.56%  (p=0.000 n=10+10)
```

### ISZERO

```
name        old time/op    new time/op    delta
OpIsZero-8     226ns ± 5%     176ns ± 5%  -22.05%  (p=0.000 n=9+10)

name        old alloc/op   new alloc/op   delta
OpIsZero-8      128B ± 0%       96B ± 0%  -25.00%  (p=0.000 n=10+10)

name        old allocs/op  new allocs/op  delta
OpIsZero-8      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=10+10)
```

### EQ


```
name     old time/op    new time/op    delta
OpEq-8      407ns ±10%     374ns ± 7%   -8.30%  (p=0.000 n=10+10)
OpEq2-8     405ns ± 9%     367ns ± 7%   -9.56%  (p=0.000 n=10+9)

name     old alloc/op   new alloc/op   delta
OpEq-8       192B ± 0%      192B ± 0%     ~     (all equal)
OpEq2-8      224B ± 0%      192B ± 0%  -14.29%  (p=0.000 n=10+10)

name     old allocs/op  new allocs/op  delta
OpEq-8       4.00 ± 0%      4.00 ± 0%     ~     (all equal)
OpEq2-8      5.00 ± 0%      4.00 ± 0%  -20.00%  (p=0.000 n=10+10)
```