@karalabe 2018-03-07
@holiman Could you rebase it so we can get it in for .4?